refactor(`SecretKeyMaterial`): Move logic to new file
