+++ src/bin/neptune-cli.rs
@@ -23,10 +23,10 @@ use neptune_cash::models::blockchain::type_scripts::neptune_coins::NeptuneCoins;
 use neptune_cash::models::state::wallet::address::KeyType;
 use neptune_cash::models::state::wallet::address::ReceivingAddress;
 use neptune_cash::models::state::wallet::coin_with_possible_timelock::CoinWithPossibleTimeLock;
+use neptune_cash::models::state::wallet::secret_key_material::SecretKeyMaterial;
 use neptune_cash::models::state::wallet::utxo_notification::PrivateNotificationData;
 use neptune_cash::models::state::wallet::utxo_notification::UtxoNotificationMedium;
 use neptune_cash::models::state::wallet::wallet_status::WalletStatus;
-use neptune_cash::models::state::wallet::SecretKeyMaterial;
 use neptune_cash::models::state::wallet::WalletSecret;
 use neptune_cash::rpc_auth;
 use neptune_cash::rpc_server::error::RpcError;
